separated AddItem and ShareList into their own components, removed console logs and renamed shareList htmlFor id for clarity, also removed sharelist form id
